+++ src/mysql/alter_paper_files_table.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `paper_files` ADD `source` INT NOT NULL DEFAULT '4' AFTER `doc_id`, ADD INDEX `source` (`source`);